@@ -227,13 +227,13 @@ class $ChannelsTable extends Channels
   }
 
   static TypeConverter<List<String>, String> $converterownCapabilities =
-      ListConverter<String>();
+      ListConverter();
   static TypeConverter<List<String>?, String?> $converterownCapabilitiesn =
       NullAwareTypeConverter.wrap($converterownCapabilities);
   static TypeConverter<Map<String, dynamic>, String> $converterconfig =
       MapConverter();
   static TypeConverter<Map<String, Object?>, String> $converterextraData =
-      MapConverter<Object?>();
+      MapConverter();
   static TypeConverter<Map<String, Object?>?, String?> $converterextraDatan =
       NullAwareTypeConverter.wrap($converterextraData);
 }
@@ -757,28 +757,42 @@ class $MessagesTable extends Messages
   late final GeneratedColumn<String> command = GeneratedColumn<String>(
       'command', aliasedName, true,
       type: DriftSqlType.string, requiredDuringInsert: false);
-  static const VerificationMeta _createdAtMeta =
-      const VerificationMeta('createdAt');
+  static const VerificationMeta _localCreatedAtMeta =
+      const VerificationMeta('localCreatedAt');
   @override
-  late final GeneratedColumn<DateTime> createdAt = GeneratedColumn<DateTime>(
-      'created_at', aliasedName, false,
-      type: DriftSqlType.dateTime,
-      requiredDuringInsert: false,
-      defaultValue: currentDateAndTime);
-  static const VerificationMeta _updatedAtMeta =
-      const VerificationMeta('updatedAt');
+  late final GeneratedColumn<DateTime> localCreatedAt =
+      GeneratedColumn<DateTime>('local_created_at', aliasedName, true,
+          type: DriftSqlType.dateTime, requiredDuringInsert: false);
+  static const VerificationMeta _remoteCreatedAtMeta =
+      const VerificationMeta('remoteCreatedAt');
   @override
-  late final GeneratedColumn<DateTime> updatedAt = GeneratedColumn<DateTime>(
-      'updated_at', aliasedName, false,
-      type: DriftSqlType.dateTime,
-      requiredDuringInsert: false,
-      defaultValue: currentDateAndTime);
-  static const VerificationMeta _deletedAtMeta =
-      const VerificationMeta('deletedAt');
+  late final GeneratedColumn<DateTime> remoteCreatedAt =
+      GeneratedColumn<DateTime>('remote_created_at', aliasedName, true,
+          type: DriftSqlType.dateTime, requiredDuringInsert: false);
+  static const VerificationMeta _localUpdatedAtMeta =
+      const VerificationMeta('localUpdatedAt');
   @override
-  late final GeneratedColumn<DateTime> deletedAt = GeneratedColumn<DateTime>(
-      'deleted_at', aliasedName, true,
-      type: DriftSqlType.dateTime, requiredDuringInsert: false);
+  late final GeneratedColumn<DateTime> localUpdatedAt =
+      GeneratedColumn<DateTime>('local_updated_at', aliasedName, true,
+          type: DriftSqlType.dateTime, requiredDuringInsert: false);
+  static const VerificationMeta _remoteUpdatedAtMeta =
+      const VerificationMeta('remoteUpdatedAt');
+  @override
+  late final GeneratedColumn<DateTime> remoteUpdatedAt =
+      GeneratedColumn<DateTime>('remote_updated_at', aliasedName, true,
+          type: DriftSqlType.dateTime, requiredDuringInsert: false);
+  static const VerificationMeta _localDeletedAtMeta =
+      const VerificationMeta('localDeletedAt');
+  @override
+  late final GeneratedColumn<DateTime> localDeletedAt =
+      GeneratedColumn<DateTime>('local_deleted_at', aliasedName, true,
+          type: DriftSqlType.dateTime, requiredDuringInsert: false);
+  static const VerificationMeta _remoteDeletedAtMeta =
+      const VerificationMeta('remoteDeletedAt');
+  @override
+  late final GeneratedColumn<DateTime> remoteDeletedAt =
+      GeneratedColumn<DateTime>('remote_deleted_at', aliasedName, true,
+          type: DriftSqlType.dateTime, requiredDuringInsert: false);
   static const VerificationMeta _userIdMeta = const VerificationMeta('userId');
   @override
   late final GeneratedColumn<String> userId = GeneratedColumn<String>(
@@ -853,9 +867,12 @@ class $MessagesTable extends Messages
         showInChannel,
         shadowed,
         command,
-        createdAt,
-        updatedAt,
-        deletedAt,
+        localCreatedAt,
+        remoteCreatedAt,
+        localUpdatedAt,
+        remoteUpdatedAt,
+        localDeletedAt,
+        remoteDeletedAt,
         userId,
         pinned,
         pinnedAt,
@@ -924,17 +941,41 @@ class $MessagesTable extends Messages
       context.handle(_commandMeta,
           command.isAcceptableOrUnknown(data['command']!, _commandMeta));
     }
-    if (data.containsKey('created_at')) {
-      context.handle(_createdAtMeta,
-          createdAt.isAcceptableOrUnknown(data['created_at']!, _createdAtMeta));
+    if (data.containsKey('local_created_at')) {
+      context.handle(
+          _localCreatedAtMeta,
+          localCreatedAt.isAcceptableOrUnknown(
+              data['local_created_at']!, _localCreatedAtMeta));
     }
-    if (data.containsKey('updated_at')) {
-      context.handle(_updatedAtMeta,
-          updatedAt.isAcceptableOrUnknown(data['updated_at']!, _updatedAtMeta));
+    if (data.containsKey('remote_created_at')) {
+      context.handle(
+          _remoteCreatedAtMeta,
+          remoteCreatedAt.isAcceptableOrUnknown(
+              data['remote_created_at']!, _remoteCreatedAtMeta));
     }
-    if (data.containsKey('deleted_at')) {
-      context.handle(_deletedAtMeta,
-          deletedAt.isAcceptableOrUnknown(data['deleted_at']!, _deletedAtMeta));
+    if (data.containsKey('local_updated_at')) {
+      context.handle(
+          _localUpdatedAtMeta,
+          localUpdatedAt.isAcceptableOrUnknown(
+              data['local_updated_at']!, _localUpdatedAtMeta));
+    }
+    if (data.containsKey('remote_updated_at')) {
+      context.handle(
+          _remoteUpdatedAtMeta,
+          remoteUpdatedAt.isAcceptableOrUnknown(
+              data['remote_updated_at']!, _remoteUpdatedAtMeta));
+    }
+    if (data.containsKey('local_deleted_at')) {
+      context.handle(
+          _localDeletedAtMeta,
+          localDeletedAt.isAcceptableOrUnknown(
+              data['local_deleted_at']!, _localDeletedAtMeta));
+    }
+    if (data.containsKey('remote_deleted_at')) {
+      context.handle(
+          _remoteDeletedAtMeta,
+          remoteDeletedAt.isAcceptableOrUnknown(
+              data['remote_deleted_at']!, _remoteDeletedAtMeta));
     }
     if (data.containsKey('user_id')) {
       context.handle(_userIdMeta,
@@ -1012,12 +1053,18 @@ class $MessagesTable extends Messages
           .read(DriftSqlType.bool, data['${effectivePrefix}shadowed'])!,
       command: attachedDatabase.typeMapping
           .read(DriftSqlType.string, data['${effectivePrefix}command']),
-      createdAt: attachedDatabase.typeMapping
-          .read(DriftSqlType.dateTime, data['${effectivePrefix}created_at'])!,
-      updatedAt: attachedDatabase.typeMapping
-          .read(DriftSqlType.dateTime, data['${effectivePrefix}updated_at'])!,
-      deletedAt: attachedDatabase.typeMapping
-          .read(DriftSqlType.dateTime, data['${effectivePrefix}deleted_at']),
+      localCreatedAt: attachedDatabase.typeMapping.read(
+          DriftSqlType.dateTime, data['${effectivePrefix}local_created_at']),
+      remoteCreatedAt: attachedDatabase.typeMapping.read(
+          DriftSqlType.dateTime, data['${effectivePrefix}remote_created_at']),
+      localUpdatedAt: attachedDatabase.typeMapping.read(
+          DriftSqlType.dateTime, data['${effectivePrefix}local_updated_at']),
+      remoteUpdatedAt: attachedDatabase.typeMapping.read(
+          DriftSqlType.dateTime, data['${effectivePrefix}remote_updated_at']),
+      localDeletedAt: attachedDatabase.typeMapping.read(
+          DriftSqlType.dateTime, data['${effectivePrefix}local_deleted_at']),
+      remoteDeletedAt: attachedDatabase.typeMapping.read(
+          DriftSqlType.dateTime, data['${effectivePrefix}remote_deleted_at']),
       userId: attachedDatabase.typeMapping
           .read(DriftSqlType.string, data['${effectivePrefix}user_id']),
       pinned: attachedDatabase.typeMapping
@@ -1044,23 +1091,23 @@ class $MessagesTable extends Messages
   }
 
   static TypeConverter<List<String>, String> $converterattachments =
-      ListConverter<String>();
+      ListConverter();
   static TypeConverter<MessageSendingStatus, int> $converterstatus =
       MessageSendingStatusConverter();
   static TypeConverter<List<String>, String> $convertermentionedUsers =
-      ListConverter<String>();
+      ListConverter();
   static TypeConverter<Map<String, int>, String> $converterreactionCounts =
-      MapConverter<int>();
+      MapConverter();
   static TypeConverter<Map<String, int>?, String?> $converterreactionCountsn =
       NullAwareTypeConverter.wrap($converterreactionCounts);
   static TypeConverter<Map<String, int>, String> $converterreactionScores =
-      MapConverter<int>();
+      MapConverter();
   static TypeConverter<Map<String, int>?, String?> $converterreactionScoresn =
       NullAwareTypeConverter.wrap($converterreactionScores);
   static TypeConverter<Map<String, String>?, String?> $converteri18n =
-      NullableMapConverter<String>();
+      NullableMapConverter();
   static TypeConverter<Map<String, Object?>, String> $converterextraData =
-      MapConverter<Object?>();
+      MapConverter();
   static TypeConverter<Map<String, Object?>?, String?> $converterextraDatan =
       NullAwareTypeConverter.wrap($converterextraData);
 }
@@ -1109,14 +1156,23 @@ class MessageEntity extends DataClass implements Insertable<MessageEntity> {
   /// A used command name.
   final String? command;
 
-  /// The DateTime when the message was created.
-  final DateTime createdAt;
+  /// The DateTime on which the message was created on the client.
+  final DateTime? localCreatedAt;
 
-  /// The DateTime when the message was updated last time.
-  final DateTime updatedAt;
+  /// The DateTime on which the message was created on the server.
+  final DateTime? remoteCreatedAt;
 
-  /// The DateTime when the message was deleted.
-  final DateTime? deletedAt;
+  /// The DateTime on which the message was updated on the client.
+  final DateTime? localUpdatedAt;
+
+  /// The DateTime on which the message was updated on the server.
+  final DateTime? remoteUpdatedAt;
+
+  /// The DateTime on which the message was deleted on the client.
+  final DateTime? localDeletedAt;
+
+  /// The DateTime on which the message was deleted on the server.
+  final DateTime? remoteDeletedAt;
 
   /// Id of the User who sent the message
   final String? userId;
@@ -1156,9 +1212,12 @@ class MessageEntity extends DataClass implements Insertable<MessageEntity> {
       this.showInChannel,
       required this.shadowed,
       this.command,
-      required this.createdAt,
-      required this.updatedAt,
-      this.deletedAt,
+      this.localCreatedAt,
+      this.remoteCreatedAt,
+      this.localUpdatedAt,
+      this.remoteUpdatedAt,
+      this.localDeletedAt,
+      this.remoteDeletedAt,
       this.userId,
       required this.pinned,
       this.pinnedAt,
@@ -1214,10 +1273,23 @@ class MessageEntity extends DataClass implements Insertable<MessageEntity> {
     if (!nullToAbsent || command != null) {
       map['command'] = Variable<String>(command);
     }
-    map['created_at'] = Variable<DateTime>(createdAt);
-    map['updated_at'] = Variable<DateTime>(updatedAt);
-    if (!nullToAbsent || deletedAt != null) {
-      map['deleted_at'] = Variable<DateTime>(deletedAt);
+    if (!nullToAbsent || localCreatedAt != null) {
+      map['local_created_at'] = Variable<DateTime>(localCreatedAt);
+    }
+    if (!nullToAbsent || remoteCreatedAt != null) {
+      map['remote_created_at'] = Variable<DateTime>(remoteCreatedAt);
+    }
+    if (!nullToAbsent || localUpdatedAt != null) {
+      map['local_updated_at'] = Variable<DateTime>(localUpdatedAt);
+    }
+    if (!nullToAbsent || remoteUpdatedAt != null) {
+      map['remote_updated_at'] = Variable<DateTime>(remoteUpdatedAt);
+    }
+    if (!nullToAbsent || localDeletedAt != null) {
+      map['local_deleted_at'] = Variable<DateTime>(localDeletedAt);
+    }
+    if (!nullToAbsent || remoteDeletedAt != null) {
+      map['remote_deleted_at'] = Variable<DateTime>(remoteDeletedAt);
     }
     if (!nullToAbsent || userId != null) {
       map['user_id'] = Variable<String>(userId);
@@ -1264,9 +1336,12 @@ class MessageEntity extends DataClass implements Insertable<MessageEntity> {
       showInChannel: serializer.fromJson<bool?>(json['showInChannel']),
       shadowed: serializer.fromJson<bool>(json['shadowed']),
       command: serializer.fromJson<String?>(json['command']),
-      createdAt: serializer.fromJson<DateTime>(json['createdAt']),
-      updatedAt: serializer.fromJson<DateTime>(json['updatedAt']),
-      deletedAt: serializer.fromJson<DateTime?>(json['deletedAt']),
+      localCreatedAt: serializer.fromJson<DateTime?>(json['localCreatedAt']),
+      remoteCreatedAt: serializer.fromJson<DateTime?>(json['remoteCreatedAt']),
+      localUpdatedAt: serializer.fromJson<DateTime?>(json['localUpdatedAt']),
+      remoteUpdatedAt: serializer.fromJson<DateTime?>(json['remoteUpdatedAt']),
+      localDeletedAt: serializer.fromJson<DateTime?>(json['localDeletedAt']),
+      remoteDeletedAt: serializer.fromJson<DateTime?>(json['remoteDeletedAt']),
       userId: serializer.fromJson<String?>(json['userId']),
       pinned: serializer.fromJson<bool>(json['pinned']),
       pinnedAt: serializer.fromJson<DateTime?>(json['pinnedAt']),
@@ -1295,9 +1370,12 @@ class MessageEntity extends DataClass implements Insertable<MessageEntity> {
       'showInChannel': serializer.toJson<bool?>(showInChannel),
       'shadowed': serializer.toJson<bool>(shadowed),
       'command': serializer.toJson<String?>(command),
-      'createdAt': serializer.toJson<DateTime>(createdAt),
-      'updatedAt': serializer.toJson<DateTime>(updatedAt),
-      'deletedAt': serializer.toJson<DateTime?>(deletedAt),
+      'localCreatedAt': serializer.toJson<DateTime?>(localCreatedAt),
+      'remoteCreatedAt': serializer.toJson<DateTime?>(remoteCreatedAt),
+      'localUpdatedAt': serializer.toJson<DateTime?>(localUpdatedAt),
+      'remoteUpdatedAt': serializer.toJson<DateTime?>(remoteUpdatedAt),
+      'localDeletedAt': serializer.toJson<DateTime?>(localDeletedAt),
+      'remoteDeletedAt': serializer.toJson<DateTime?>(remoteDeletedAt),
       'userId': serializer.toJson<String?>(userId),
       'pinned': serializer.toJson<bool>(pinned),
       'pinnedAt': serializer.toJson<DateTime?>(pinnedAt),
@@ -1324,9 +1402,12 @@ class MessageEntity extends DataClass implements Insertable<MessageEntity> {
           Value<bool?> showInChannel = const Value.absent(),
           bool? shadowed,
           Value<String?> command = const Value.absent(),
-          DateTime? createdAt,
-          DateTime? updatedAt,
-          Value<DateTime?> deletedAt = const Value.absent(),
+          Value<DateTime?> localCreatedAt = const Value.absent(),
+          Value<DateTime?> remoteCreatedAt = const Value.absent(),
+          Value<DateTime?> localUpdatedAt = const Value.absent(),
+          Value<DateTime?> remoteUpdatedAt = const Value.absent(),
+          Value<DateTime?> localDeletedAt = const Value.absent(),
+          Value<DateTime?> remoteDeletedAt = const Value.absent(),
           Value<String?> userId = const Value.absent(),
           bool? pinned,
           Value<DateTime?> pinnedAt = const Value.absent(),
@@ -1355,9 +1436,21 @@ class MessageEntity extends DataClass implements Insertable<MessageEntity> {
             showInChannel.present ? showInChannel.value : this.showInChannel,
         shadowed: shadowed ?? this.shadowed,
         command: command.present ? command.value : this.command,
-        createdAt: createdAt ?? this.createdAt,
-        updatedAt: updatedAt ?? this.updatedAt,
-        deletedAt: deletedAt.present ? deletedAt.value : this.deletedAt,
+        localCreatedAt:
+            localCreatedAt.present ? localCreatedAt.value : this.localCreatedAt,
+        remoteCreatedAt: remoteCreatedAt.present
+            ? remoteCreatedAt.value
+            : this.remoteCreatedAt,
+        localUpdatedAt:
+            localUpdatedAt.present ? localUpdatedAt.value : this.localUpdatedAt,
+        remoteUpdatedAt: remoteUpdatedAt.present
+            ? remoteUpdatedAt.value
+            : this.remoteUpdatedAt,
+        localDeletedAt:
+            localDeletedAt.present ? localDeletedAt.value : this.localDeletedAt,
+        remoteDeletedAt: remoteDeletedAt.present
+            ? remoteDeletedAt.value
+            : this.remoteDeletedAt,
         userId: userId.present ? userId.value : this.userId,
         pinned: pinned ?? this.pinned,
         pinnedAt: pinnedAt.present ? pinnedAt.value : this.pinnedAt,
@@ -1385,9 +1478,12 @@ class MessageEntity extends DataClass implements Insertable<MessageEntity> {
           ..write('showInChannel: $showInChannel, ')
           ..write('shadowed: $shadowed, ')
           ..write('command: $command, ')
-          ..write('createdAt: $createdAt, ')
-          ..write('updatedAt: $updatedAt, ')
-          ..write('deletedAt: $deletedAt, ')
+          ..write('localCreatedAt: $localCreatedAt, ')
+          ..write('remoteCreatedAt: $remoteCreatedAt, ')
+          ..write('localUpdatedAt: $localUpdatedAt, ')
+          ..write('remoteUpdatedAt: $remoteUpdatedAt, ')
+          ..write('localDeletedAt: $localDeletedAt, ')
+          ..write('remoteDeletedAt: $remoteDeletedAt, ')
           ..write('userId: $userId, ')
           ..write('pinned: $pinned, ')
           ..write('pinnedAt: $pinnedAt, ')
@@ -1416,9 +1512,12 @@ class MessageEntity extends DataClass implements Insertable<MessageEntity> {
         showInChannel,
         shadowed,
         command,
-        createdAt,
-        updatedAt,
-        deletedAt,
+        localCreatedAt,
+        remoteCreatedAt,
+        localUpdatedAt,
+        remoteUpdatedAt,
+        localDeletedAt,
+        remoteDeletedAt,
         userId,
         pinned,
         pinnedAt,
@@ -1446,9 +1545,12 @@ class MessageEntity extends DataClass implements Insertable<MessageEntity> {
           other.showInChannel == this.showInChannel &&
           other.shadowed == this.shadowed &&
           other.command == this.command &&
-          other.createdAt == this.createdAt &&
-          other.updatedAt == this.updatedAt &&
-          other.deletedAt == this.deletedAt &&
+          other.localCreatedAt == this.localCreatedAt &&
+          other.remoteCreatedAt == this.remoteCreatedAt &&
+          other.localUpdatedAt == this.localUpdatedAt &&
+          other.remoteUpdatedAt == this.remoteUpdatedAt &&
+          other.localDeletedAt == this.localDeletedAt &&
+          other.remoteDeletedAt == this.remoteDeletedAt &&
           other.userId == this.userId &&
           other.pinned == this.pinned &&
           other.pinnedAt == this.pinnedAt &&
@@ -1474,9 +1576,12 @@ class MessagesCompanion extends UpdateCompanion<MessageEntity> {
   final Value<bool?> showInChannel;
   final Value<bool> shadowed;
   final Value<String?> command;
-  final Value<DateTime> createdAt;
-  final Value<DateTime> updatedAt;
-  final Value<DateTime?> deletedAt;
+  final Value<DateTime?> localCreatedAt;
+  final Value<DateTime?> remoteCreatedAt;
+  final Value<DateTime?> localUpdatedAt;
+  final Value<DateTime?> remoteUpdatedAt;
+  final Value<DateTime?> localDeletedAt;
+  final Value<DateTime?> remoteDeletedAt;
   final Value<String?> userId;
   final Value<bool> pinned;
   final Value<DateTime?> pinnedAt;
@@ -1501,9 +1606,12 @@ class MessagesCompanion extends UpdateCompanion<MessageEntity> {
     this.showInChannel = const Value.absent(),
     this.shadowed = const Value.absent(),
     this.command = const Value.absent(),
-    this.createdAt = const Value.absent(),
-    this.updatedAt = const Value.absent(),
-    this.deletedAt = const Value.absent(),
+    this.localCreatedAt = const Value.absent(),
+    this.remoteCreatedAt = const Value.absent(),
+    this.localUpdatedAt = const Value.absent(),
+    this.remoteUpdatedAt = const Value.absent(),
+    this.localDeletedAt = const Value.absent(),
+    this.remoteDeletedAt = const Value.absent(),
     this.userId = const Value.absent(),
     this.pinned = const Value.absent(),
     this.pinnedAt = const Value.absent(),
@@ -1529,9 +1637,12 @@ class MessagesCompanion extends UpdateCompanion<MessageEntity> {
     this.showInChannel = const Value.absent(),
     this.shadowed = const Value.absent(),
     this.command = const Value.absent(),
-    this.createdAt = const Value.absent(),
-    this.updatedAt = const Value.absent(),
-    this.deletedAt = const Value.absent(),
+    this.localCreatedAt = const Value.absent(),
+    this.remoteCreatedAt = const Value.absent(),
+    this.localUpdatedAt = const Value.absent(),
+    this.remoteUpdatedAt = const Value.absent(),
+    this.localDeletedAt = const Value.absent(),
+    this.remoteDeletedAt = const Value.absent(),
     this.userId = const Value.absent(),
     this.pinned = const Value.absent(),
     this.pinnedAt = const Value.absent(),
@@ -1560,9 +1671,12 @@ class MessagesCompanion extends UpdateCompanion<MessageEntity> {
     Expression<bool>? showInChannel,
     Expression<bool>? shadowed,
     Expression<String>? command,
-    Expression<DateTime>? createdAt,
-    Expression<DateTime>? updatedAt,
-    Expression<DateTime>? deletedAt,
+    Expression<DateTime>? localCreatedAt,
+    Expression<DateTime>? remoteCreatedAt,
+    Expression<DateTime>? localUpdatedAt,
+    Expression<DateTime>? remoteUpdatedAt,
+    Expression<DateTime>? localDeletedAt,
+    Expression<DateTime>? remoteDeletedAt,
     Expression<String>? userId,
     Expression<bool>? pinned,
     Expression<DateTime>? pinnedAt,
@@ -1588,9 +1702,12 @@ class MessagesCompanion extends UpdateCompanion<MessageEntity> {
       if (showInChannel != null) 'show_in_channel': showInChannel,
       if (shadowed != null) 'shadowed': shadowed,
       if (command != null) 'command': command,
-      if (createdAt != null) 'created_at': createdAt,
-      if (updatedAt != null) 'updated_at': updatedAt,
-      if (deletedAt != null) 'deleted_at': deletedAt,
+      if (localCreatedAt != null) 'local_created_at': localCreatedAt,
+      if (remoteCreatedAt != null) 'remote_created_at': remoteCreatedAt,
+      if (localUpdatedAt != null) 'local_updated_at': localUpdatedAt,
+      if (remoteUpdatedAt != null) 'remote_updated_at': remoteUpdatedAt,
+      if (localDeletedAt != null) 'local_deleted_at': localDeletedAt,
+      if (remoteDeletedAt != null) 'remote_deleted_at': remoteDeletedAt,
       if (userId != null) 'user_id': userId,
       if (pinned != null) 'pinned': pinned,
       if (pinnedAt != null) 'pinned_at': pinnedAt,
@@ -1618,9 +1735,12 @@ class MessagesCompanion extends UpdateCompanion<MessageEntity> {
       Value<bool?>? showInChannel,
       Value<bool>? shadowed,
       Value<String?>? command,
-      Value<DateTime>? createdAt,
-      Value<DateTime>? updatedAt,
-      Value<DateTime?>? deletedAt,
+      Value<DateTime?>? localCreatedAt,
+      Value<DateTime?>? remoteCreatedAt,
+      Value<DateTime?>? localUpdatedAt,
+      Value<DateTime?>? remoteUpdatedAt,
+      Value<DateTime?>? localDeletedAt,
+      Value<DateTime?>? remoteDeletedAt,
       Value<String?>? userId,
       Value<bool>? pinned,
       Value<DateTime?>? pinnedAt,
@@ -1645,9 +1765,12 @@ class MessagesCompanion extends UpdateCompanion<MessageEntity> {
       showInChannel: showInChannel ?? this.showInChannel,
       shadowed: shadowed ?? this.shadowed,
       command: command ?? this.command,
-      createdAt: createdAt ?? this.createdAt,
-      updatedAt: updatedAt ?? this.updatedAt,
-      deletedAt: deletedAt ?? this.deletedAt,
+      localCreatedAt: localCreatedAt ?? this.localCreatedAt,
+      remoteCreatedAt: remoteCreatedAt ?? this.remoteCreatedAt,
+      localUpdatedAt: localUpdatedAt ?? this.localUpdatedAt,
+      remoteUpdatedAt: remoteUpdatedAt ?? this.remoteUpdatedAt,
+      localDeletedAt: localDeletedAt ?? this.localDeletedAt,
+      remoteDeletedAt: remoteDeletedAt ?? this.remoteDeletedAt,
       userId: userId ?? this.userId,
       pinned: pinned ?? this.pinned,
       pinnedAt: pinnedAt ?? this.pinnedAt,
@@ -1713,14 +1836,23 @@ class MessagesCompanion extends UpdateCompanion<MessageEntity> {
     if (command.present) {
       map['command'] = Variable<String>(command.value);
     }
-    if (createdAt.present) {
-      map['created_at'] = Variable<DateTime>(createdAt.value);
+    if (localCreatedAt.present) {
+      map['local_created_at'] = Variable<DateTime>(localCreatedAt.value);
     }
-    if (updatedAt.present) {
-      map['updated_at'] = Variable<DateTime>(updatedAt.value);
+    if (remoteCreatedAt.present) {
+      map['remote_created_at'] = Variable<DateTime>(remoteCreatedAt.value);
     }
-    if (deletedAt.present) {
-      map['deleted_at'] = Variable<DateTime>(deletedAt.value);
+    if (localUpdatedAt.present) {
+      map['local_updated_at'] = Variable<DateTime>(localUpdatedAt.value);
+    }
+    if (remoteUpdatedAt.present) {
+      map['remote_updated_at'] = Variable<DateTime>(remoteUpdatedAt.value);
+    }
+    if (localDeletedAt.present) {
+      map['local_deleted_at'] = Variable<DateTime>(localDeletedAt.value);
+    }
+    if (remoteDeletedAt.present) {
+      map['remote_deleted_at'] = Variable<DateTime>(remoteDeletedAt.value);
     }
     if (userId.present) {
       map['user_id'] = Variable<String>(userId.value);
@@ -1771,9 +1903,12 @@ class MessagesCompanion extends UpdateCompanion<MessageEntity> {
           ..write('showInChannel: $showInChannel, ')
           ..write('shadowed: $shadowed, ')
           ..write('command: $command, ')
-          ..write('createdAt: $createdAt, ')
-          ..write('updatedAt: $updatedAt, ')
-          ..write('deletedAt: $deletedAt, ')
+          ..write('localCreatedAt: $localCreatedAt, ')
+          ..write('remoteCreatedAt: $remoteCreatedAt, ')
+          ..write('localUpdatedAt: $localUpdatedAt, ')
+          ..write('remoteUpdatedAt: $remoteUpdatedAt, ')
+          ..write('localDeletedAt: $localDeletedAt, ')
+          ..write('remoteDeletedAt: $remoteDeletedAt, ')
           ..write('userId: $userId, ')
           ..write('pinned: $pinned, ')
           ..write('pinnedAt: $pinnedAt, ')
@@ -1905,28 +2040,42 @@ class $PinnedMessagesTable extends PinnedMessages
   late final GeneratedColumn<String> command = GeneratedColumn<String>(
       'command', aliasedName, true,
       type: DriftSqlType.string, requiredDuringInsert: false);
-  static const VerificationMeta _createdAtMeta =
-      const VerificationMeta('createdAt');
+  static const VerificationMeta _localCreatedAtMeta =
+      const VerificationMeta('localCreatedAt');
   @override
-  late final GeneratedColumn<DateTime> createdAt = GeneratedColumn<DateTime>(
-      'created_at', aliasedName, false,
-      type: DriftSqlType.dateTime,
-      requiredDuringInsert: false,
-      defaultValue: currentDateAndTime);
-  static const VerificationMeta _updatedAtMeta =
-      const VerificationMeta('updatedAt');
+  late final GeneratedColumn<DateTime> localCreatedAt =
+      GeneratedColumn<DateTime>('local_created_at', aliasedName, true,
+          type: DriftSqlType.dateTime, requiredDuringInsert: false);
+  static const VerificationMeta _remoteCreatedAtMeta =
+      const VerificationMeta('remoteCreatedAt');
   @override
-  late final GeneratedColumn<DateTime> updatedAt = GeneratedColumn<DateTime>(
-      'updated_at', aliasedName, false,
-      type: DriftSqlType.dateTime,
-      requiredDuringInsert: false,
-      defaultValue: currentDateAndTime);
-  static const VerificationMeta _deletedAtMeta =
-      const VerificationMeta('deletedAt');
+  late final GeneratedColumn<DateTime> remoteCreatedAt =
+      GeneratedColumn<DateTime>('remote_created_at', aliasedName, true,
+          type: DriftSqlType.dateTime, requiredDuringInsert: false);
+  static const VerificationMeta _localUpdatedAtMeta =
+      const VerificationMeta('localUpdatedAt');
   @override
-  late final GeneratedColumn<DateTime> deletedAt = GeneratedColumn<DateTime>(
-      'deleted_at', aliasedName, true,
-      type: DriftSqlType.dateTime, requiredDuringInsert: false);
+  late final GeneratedColumn<DateTime> localUpdatedAt =
+      GeneratedColumn<DateTime>('local_updated_at', aliasedName, true,
+          type: DriftSqlType.dateTime, requiredDuringInsert: false);
+  static const VerificationMeta _remoteUpdatedAtMeta =
+      const VerificationMeta('remoteUpdatedAt');
+  @override
+  late final GeneratedColumn<DateTime> remoteUpdatedAt =
+      GeneratedColumn<DateTime>('remote_updated_at', aliasedName, true,
+          type: DriftSqlType.dateTime, requiredDuringInsert: false);
+  static const VerificationMeta _localDeletedAtMeta =
+      const VerificationMeta('localDeletedAt');
+  @override
+  late final GeneratedColumn<DateTime> localDeletedAt =
+      GeneratedColumn<DateTime>('local_deleted_at', aliasedName, true,
+          type: DriftSqlType.dateTime, requiredDuringInsert: false);
+  static const VerificationMeta _remoteDeletedAtMeta =
+      const VerificationMeta('remoteDeletedAt');
+  @override
+  late final GeneratedColumn<DateTime> remoteDeletedAt =
+      GeneratedColumn<DateTime>('remote_deleted_at', aliasedName, true,
+          type: DriftSqlType.dateTime, requiredDuringInsert: false);
   static const VerificationMeta _userIdMeta = const VerificationMeta('userId');
   @override
   late final GeneratedColumn<String> userId = GeneratedColumn<String>(
@@ -2002,9 +2151,12 @@ class $PinnedMessagesTable extends PinnedMessages
         showInChannel,
         shadowed,
         command,
-        createdAt,
-        updatedAt,
-        deletedAt,
+        localCreatedAt,
+        remoteCreatedAt,
+        localUpdatedAt,
+        remoteUpdatedAt,
+        localDeletedAt,
+        remoteDeletedAt,
         userId,
         pinned,
         pinnedAt,
@@ -2074,17 +2226,41 @@ class $PinnedMessagesTable extends PinnedMessages
       context.handle(_commandMeta,
           command.isAcceptableOrUnknown(data['command']!, _commandMeta));
     }
-    if (data.containsKey('created_at')) {
-      context.handle(_createdAtMeta,
-          createdAt.isAcceptableOrUnknown(data['created_at']!, _createdAtMeta));
+    if (data.containsKey('local_created_at')) {
+      context.handle(
+          _localCreatedAtMeta,
+          localCreatedAt.isAcceptableOrUnknown(
+              data['local_created_at']!, _localCreatedAtMeta));
     }
-    if (data.containsKey('updated_at')) {
-      context.handle(_updatedAtMeta,
-          updatedAt.isAcceptableOrUnknown(data['updated_at']!, _updatedAtMeta));
+    if (data.containsKey('remote_created_at')) {
+      context.handle(
+          _remoteCreatedAtMeta,
+          remoteCreatedAt.isAcceptableOrUnknown(
+              data['remote_created_at']!, _remoteCreatedAtMeta));
     }
-    if (data.containsKey('deleted_at')) {
-      context.handle(_deletedAtMeta,
-          deletedAt.isAcceptableOrUnknown(data['deleted_at']!, _deletedAtMeta));
+    if (data.containsKey('local_updated_at')) {
+      context.handle(
+          _localUpdatedAtMeta,
+          localUpdatedAt.isAcceptableOrUnknown(
+              data['local_updated_at']!, _localUpdatedAtMeta));
+    }
+    if (data.containsKey('remote_updated_at')) {
+      context.handle(
+          _remoteUpdatedAtMeta,
+          remoteUpdatedAt.isAcceptableOrUnknown(
+              data['remote_updated_at']!, _remoteUpdatedAtMeta));
+    }
+    if (data.containsKey('local_deleted_at')) {
+      context.handle(
+          _localDeletedAtMeta,
+          localDeletedAt.isAcceptableOrUnknown(
+              data['local_deleted_at']!, _localDeletedAtMeta));
+    }
+    if (data.containsKey('remote_deleted_at')) {
+      context.handle(
+          _remoteDeletedAtMeta,
+          remoteDeletedAt.isAcceptableOrUnknown(
+              data['remote_deleted_at']!, _remoteDeletedAtMeta));
     }
     if (data.containsKey('user_id')) {
       context.handle(_userIdMeta,
@@ -2162,12 +2338,18 @@ class $PinnedMessagesTable extends PinnedMessages
           .read(DriftSqlType.bool, data['${effectivePrefix}shadowed'])!,
       command: attachedDatabase.typeMapping
           .read(DriftSqlType.string, data['${effectivePrefix}command']),
-      createdAt: attachedDatabase.typeMapping
-          .read(DriftSqlType.dateTime, data['${effectivePrefix}created_at'])!,
-      updatedAt: attachedDatabase.typeMapping
-          .read(DriftSqlType.dateTime, data['${effectivePrefix}updated_at'])!,
-      deletedAt: attachedDatabase.typeMapping
-          .read(DriftSqlType.dateTime, data['${effectivePrefix}deleted_at']),
+      localCreatedAt: attachedDatabase.typeMapping.read(
+          DriftSqlType.dateTime, data['${effectivePrefix}local_created_at']),
+      remoteCreatedAt: attachedDatabase.typeMapping.read(
+          DriftSqlType.dateTime, data['${effectivePrefix}remote_created_at']),
+      localUpdatedAt: attachedDatabase.typeMapping.read(
+          DriftSqlType.dateTime, data['${effectivePrefix}local_updated_at']),
+      remoteUpdatedAt: attachedDatabase.typeMapping.read(
+          DriftSqlType.dateTime, data['${effectivePrefix}remote_updated_at']),
+      localDeletedAt: attachedDatabase.typeMapping.read(
+          DriftSqlType.dateTime, data['${effectivePrefix}local_deleted_at']),
+      remoteDeletedAt: attachedDatabase.typeMapping.read(
+          DriftSqlType.dateTime, data['${effectivePrefix}remote_deleted_at']),
       userId: attachedDatabase.typeMapping
           .read(DriftSqlType.string, data['${effectivePrefix}user_id']),
       pinned: attachedDatabase.typeMapping
@@ -2195,23 +2377,23 @@ class $PinnedMessagesTable extends PinnedMessages
   }
 
   static TypeConverter<List<String>, String> $converterattachments =
-      ListConverter<String>();
+      ListConverter();
   static TypeConverter<MessageSendingStatus, int> $converterstatus =
       MessageSendingStatusConverter();
   static TypeConverter<List<String>, String> $convertermentionedUsers =
-      ListConverter<String>();
+      ListConverter();
   static TypeConverter<Map<String, int>, String> $converterreactionCounts =
-      MapConverter<int>();
+      MapConverter();
   static TypeConverter<Map<String, int>?, String?> $converterreactionCountsn =
       NullAwareTypeConverter.wrap($converterreactionCounts);
   static TypeConverter<Map<String, int>, String> $converterreactionScores =
-      MapConverter<int>();
+      MapConverter();
   static TypeConverter<Map<String, int>?, String?> $converterreactionScoresn =
       NullAwareTypeConverter.wrap($converterreactionScores);
   static TypeConverter<Map<String, String>?, String?> $converteri18n =
-      NullableMapConverter<String>();
+      NullableMapConverter();
   static TypeConverter<Map<String, Object?>, String> $converterextraData =
-      MapConverter<Object?>();
+      MapConverter();
   static TypeConverter<Map<String, Object?>?, String?> $converterextraDatan =
       NullAwareTypeConverter.wrap($converterextraData);
 }
@@ -2261,14 +2443,23 @@ class PinnedMessageEntity extends DataClass
   /// A used command name.
   final String? command;
 
-  /// The DateTime when the message was created.
-  final DateTime createdAt;
+  /// The DateTime on which the message was created on the client.
+  final DateTime? localCreatedAt;
 
-  /// The DateTime when the message was updated last time.
-  final DateTime updatedAt;
+  /// The DateTime on which the message was created on the server.
+  final DateTime? remoteCreatedAt;
 
-  /// The DateTime when the message was deleted.
-  final DateTime? deletedAt;
+  /// The DateTime on which the message was updated on the client.
+  final DateTime? localUpdatedAt;
+
+  /// The DateTime on which the message was updated on the server.
+  final DateTime? remoteUpdatedAt;
+
+  /// The DateTime on which the message was deleted on the client.
+  final DateTime? localDeletedAt;
+
+  /// The DateTime on which the message was deleted on the server.
+  final DateTime? remoteDeletedAt;
 
   /// Id of the User who sent the message
   final String? userId;
@@ -2308,9 +2499,12 @@ class PinnedMessageEntity extends DataClass
       this.showInChannel,
       required this.shadowed,
       this.command,
-      required this.createdAt,
-      required this.updatedAt,
-      this.deletedAt,
+      this.localCreatedAt,
+      this.remoteCreatedAt,
+      this.localUpdatedAt,
+      this.remoteUpdatedAt,
+      this.localDeletedAt,
+      this.remoteDeletedAt,
       this.userId,
       required this.pinned,
       this.pinnedAt,
@@ -2366,10 +2560,23 @@ class PinnedMessageEntity extends DataClass
     if (!nullToAbsent || command != null) {
       map['command'] = Variable<String>(command);
     }
-    map['created_at'] = Variable<DateTime>(createdAt);
-    map['updated_at'] = Variable<DateTime>(updatedAt);
-    if (!nullToAbsent || deletedAt != null) {
-      map['deleted_at'] = Variable<DateTime>(deletedAt);
+    if (!nullToAbsent || localCreatedAt != null) {
+      map['local_created_at'] = Variable<DateTime>(localCreatedAt);
+    }
+    if (!nullToAbsent || remoteCreatedAt != null) {
+      map['remote_created_at'] = Variable<DateTime>(remoteCreatedAt);
+    }
+    if (!nullToAbsent || localUpdatedAt != null) {
+      map['local_updated_at'] = Variable<DateTime>(localUpdatedAt);
+    }
+    if (!nullToAbsent || remoteUpdatedAt != null) {
+      map['remote_updated_at'] = Variable<DateTime>(remoteUpdatedAt);
+    }
+    if (!nullToAbsent || localDeletedAt != null) {
+      map['local_deleted_at'] = Variable<DateTime>(localDeletedAt);
+    }
+    if (!nullToAbsent || remoteDeletedAt != null) {
+      map['remote_deleted_at'] = Variable<DateTime>(remoteDeletedAt);
     }
     if (!nullToAbsent || userId != null) {
       map['user_id'] = Variable<String>(userId);
@@ -2416,9 +2623,12 @@ class PinnedMessageEntity extends DataClass
       showInChannel: serializer.fromJson<bool?>(json['showInChannel']),
       shadowed: serializer.fromJson<bool>(json['shadowed']),
       command: serializer.fromJson<String?>(json['command']),
-      createdAt: serializer.fromJson<DateTime>(json['createdAt']),
-      updatedAt: serializer.fromJson<DateTime>(json['updatedAt']),
-      deletedAt: serializer.fromJson<DateTime?>(json['deletedAt']),
+      localCreatedAt: serializer.fromJson<DateTime?>(json['localCreatedAt']),
+      remoteCreatedAt: serializer.fromJson<DateTime?>(json['remoteCreatedAt']),
+      localUpdatedAt: serializer.fromJson<DateTime?>(json['localUpdatedAt']),
+      remoteUpdatedAt: serializer.fromJson<DateTime?>(json['remoteUpdatedAt']),
+      localDeletedAt: serializer.fromJson<DateTime?>(json['localDeletedAt']),
+      remoteDeletedAt: serializer.fromJson<DateTime?>(json['remoteDeletedAt']),
       userId: serializer.fromJson<String?>(json['userId']),
       pinned: serializer.fromJson<bool>(json['pinned']),
       pinnedAt: serializer.fromJson<DateTime?>(json['pinnedAt']),
@@ -2447,9 +2657,12 @@ class PinnedMessageEntity extends DataClass
       'showInChannel': serializer.toJson<bool?>(showInChannel),
       'shadowed': serializer.toJson<bool>(shadowed),
       'command': serializer.toJson<String?>(command),
-      'createdAt': serializer.toJson<DateTime>(createdAt),
-      'updatedAt': serializer.toJson<DateTime>(updatedAt),
-      'deletedAt': serializer.toJson<DateTime?>(deletedAt),
+      'localCreatedAt': serializer.toJson<DateTime?>(localCreatedAt),
+      'remoteCreatedAt': serializer.toJson<DateTime?>(remoteCreatedAt),
+      'localUpdatedAt': serializer.toJson<DateTime?>(localUpdatedAt),
+      'remoteUpdatedAt': serializer.toJson<DateTime?>(remoteUpdatedAt),
+      'localDeletedAt': serializer.toJson<DateTime?>(localDeletedAt),
+      'remoteDeletedAt': serializer.toJson<DateTime?>(remoteDeletedAt),
       'userId': serializer.toJson<String?>(userId),
       'pinned': serializer.toJson<bool>(pinned),
       'pinnedAt': serializer.toJson<DateTime?>(pinnedAt),
@@ -2476,9 +2689,12 @@ class PinnedMessageEntity extends DataClass
           Value<bool?> showInChannel = const Value.absent(),
           bool? shadowed,
           Value<String?> command = const Value.absent(),
-          DateTime? createdAt,
-          DateTime? updatedAt,
-          Value<DateTime?> deletedAt = const Value.absent(),
+          Value<DateTime?> localCreatedAt = const Value.absent(),
+          Value<DateTime?> remoteCreatedAt = const Value.absent(),
+          Value<DateTime?> localUpdatedAt = const Value.absent(),
+          Value<DateTime?> remoteUpdatedAt = const Value.absent(),
+          Value<DateTime?> localDeletedAt = const Value.absent(),
+          Value<DateTime?> remoteDeletedAt = const Value.absent(),
           Value<String?> userId = const Value.absent(),
           bool? pinned,
           Value<DateTime?> pinnedAt = const Value.absent(),
@@ -2507,9 +2723,21 @@ class PinnedMessageEntity extends DataClass
             showInChannel.present ? showInChannel.value : this.showInChannel,
         shadowed: shadowed ?? this.shadowed,
         command: command.present ? command.value : this.command,
-        createdAt: createdAt ?? this.createdAt,
-        updatedAt: updatedAt ?? this.updatedAt,
-        deletedAt: deletedAt.present ? deletedAt.value : this.deletedAt,
+        localCreatedAt:
+            localCreatedAt.present ? localCreatedAt.value : this.localCreatedAt,
+        remoteCreatedAt: remoteCreatedAt.present
+            ? remoteCreatedAt.value
+            : this.remoteCreatedAt,
+        localUpdatedAt:
+            localUpdatedAt.present ? localUpdatedAt.value : this.localUpdatedAt,
+        remoteUpdatedAt: remoteUpdatedAt.present
+            ? remoteUpdatedAt.value
+            : this.remoteUpdatedAt,
+        localDeletedAt:
+            localDeletedAt.present ? localDeletedAt.value : this.localDeletedAt,
+        remoteDeletedAt: remoteDeletedAt.present
+            ? remoteDeletedAt.value
+            : this.remoteDeletedAt,
         userId: userId.present ? userId.value : this.userId,
         pinned: pinned ?? this.pinned,
         pinnedAt: pinnedAt.present ? pinnedAt.value : this.pinnedAt,
@@ -2537,9 +2765,12 @@ class PinnedMessageEntity extends DataClass
           ..write('showInChannel: $showInChannel, ')
           ..write('shadowed: $shadowed, ')
           ..write('command: $command, ')
-          ..write('createdAt: $createdAt, ')
-          ..write('updatedAt: $updatedAt, ')
-          ..write('deletedAt: $deletedAt, ')
+          ..write('localCreatedAt: $localCreatedAt, ')
+          ..write('remoteCreatedAt: $remoteCreatedAt, ')
+          ..write('localUpdatedAt: $localUpdatedAt, ')
+          ..write('remoteUpdatedAt: $remoteUpdatedAt, ')
+          ..write('localDeletedAt: $localDeletedAt, ')
+          ..write('remoteDeletedAt: $remoteDeletedAt, ')
           ..write('userId: $userId, ')
           ..write('pinned: $pinned, ')
           ..write('pinnedAt: $pinnedAt, ')
@@ -2568,9 +2799,12 @@ class PinnedMessageEntity extends DataClass
         showInChannel,
         shadowed,
         command,
-        createdAt,
-        updatedAt,
-        deletedAt,
+        localCreatedAt,
+        remoteCreatedAt,
+        localUpdatedAt,
+        remoteUpdatedAt,
+        localDeletedAt,
+        remoteDeletedAt,
         userId,
         pinned,
         pinnedAt,
@@ -2598,9 +2832,12 @@ class PinnedMessageEntity extends DataClass
           other.showInChannel == this.showInChannel &&
           other.shadowed == this.shadowed &&
           other.command == this.command &&
-          other.createdAt == this.createdAt &&
-          other.updatedAt == this.updatedAt &&
-          other.deletedAt == this.deletedAt &&
+          other.localCreatedAt == this.localCreatedAt &&
+          other.remoteCreatedAt == this.remoteCreatedAt &&
+          other.localUpdatedAt == this.localUpdatedAt &&
+          other.remoteUpdatedAt == this.remoteUpdatedAt &&
+          other.localDeletedAt == this.localDeletedAt &&
+          other.remoteDeletedAt == this.remoteDeletedAt &&
           other.userId == this.userId &&
           other.pinned == this.pinned &&
           other.pinnedAt == this.pinnedAt &&
@@ -2626,9 +2863,12 @@ class PinnedMessagesCompanion extends UpdateCompanion<PinnedMessageEntity> {
   final Value<bool?> showInChannel;
   final Value<bool> shadowed;
   final Value<String?> command;
-  final Value<DateTime> createdAt;
-  final Value<DateTime> updatedAt;
-  final Value<DateTime?> deletedAt;
+  final Value<DateTime?> localCreatedAt;
+  final Value<DateTime?> remoteCreatedAt;
+  final Value<DateTime?> localUpdatedAt;
+  final Value<DateTime?> remoteUpdatedAt;
+  final Value<DateTime?> localDeletedAt;
+  final Value<DateTime?> remoteDeletedAt;
   final Value<String?> userId;
   final Value<bool> pinned;
   final Value<DateTime?> pinnedAt;
@@ -2653,9 +2893,12 @@ class PinnedMessagesCompanion extends UpdateCompanion<PinnedMessageEntity> {
     this.showInChannel = const Value.absent(),
     this.shadowed = const Value.absent(),
     this.command = const Value.absent(),
-    this.createdAt = const Value.absent(),
-    this.updatedAt = const Value.absent(),
-    this.deletedAt = const Value.absent(),
+    this.localCreatedAt = const Value.absent(),
+    this.remoteCreatedAt = const Value.absent(),
+    this.localUpdatedAt = const Value.absent(),
+    this.remoteUpdatedAt = const Value.absent(),
+    this.localDeletedAt = const Value.absent(),
+    this.remoteDeletedAt = const Value.absent(),
     this.userId = const Value.absent(),
     this.pinned = const Value.absent(),
     this.pinnedAt = const Value.absent(),
@@ -2681,9 +2924,12 @@ class PinnedMessagesCompanion extends UpdateCompanion<PinnedMessageEntity> {
     this.showInChannel = const Value.absent(),
     this.shadowed = const Value.absent(),
     this.command = const Value.absent(),
-    this.createdAt = const Value.absent(),
-    this.updatedAt = const Value.absent(),
-    this.deletedAt = const Value.absent(),
+    this.localCreatedAt = const Value.absent(),
+    this.remoteCreatedAt = const Value.absent(),
+    this.localUpdatedAt = const Value.absent(),
+    this.remoteUpdatedAt = const Value.absent(),
+    this.localDeletedAt = const Value.absent(),
+    this.remoteDeletedAt = const Value.absent(),
     this.userId = const Value.absent(),
     this.pinned = const Value.absent(),
     this.pinnedAt = const Value.absent(),
@@ -2712,9 +2958,12 @@ class PinnedMessagesCompanion extends UpdateCompanion<PinnedMessageEntity> {
     Expression<bool>? showInChannel,
     Expression<bool>? shadowed,
     Expression<String>? command,
-    Expression<DateTime>? createdAt,
-    Expression<DateTime>? updatedAt,
-    Expression<DateTime>? deletedAt,
+    Expression<DateTime>? localCreatedAt,
+    Expression<DateTime>? remoteCreatedAt,
+    Expression<DateTime>? localUpdatedAt,
+    Expression<DateTime>? remoteUpdatedAt,
+    Expression<DateTime>? localDeletedAt,
+    Expression<DateTime>? remoteDeletedAt,
     Expression<String>? userId,
     Expression<bool>? pinned,
     Expression<DateTime>? pinnedAt,
@@ -2740,9 +2989,12 @@ class PinnedMessagesCompanion extends UpdateCompanion<PinnedMessageEntity> {
       if (showInChannel != null) 'show_in_channel': showInChannel,
       if (shadowed != null) 'shadowed': shadowed,
       if (command != null) 'command': command,
-      if (createdAt != null) 'created_at': createdAt,
-      if (updatedAt != null) 'updated_at': updatedAt,
-      if (deletedAt != null) 'deleted_at': deletedAt,
+      if (localCreatedAt != null) 'local_created_at': localCreatedAt,
+      if (remoteCreatedAt != null) 'remote_created_at': remoteCreatedAt,
+      if (localUpdatedAt != null) 'local_updated_at': localUpdatedAt,
+      if (remoteUpdatedAt != null) 'remote_updated_at': remoteUpdatedAt,
+      if (localDeletedAt != null) 'local_deleted_at': localDeletedAt,
+      if (remoteDeletedAt != null) 'remote_deleted_at': remoteDeletedAt,
       if (userId != null) 'user_id': userId,
       if (pinned != null) 'pinned': pinned,
       if (pinnedAt != null) 'pinned_at': pinnedAt,
@@ -2770,9 +3022,12 @@ class PinnedMessagesCompanion extends UpdateCompanion<PinnedMessageEntity> {
       Value<bool?>? showInChannel,
       Value<bool>? shadowed,
       Value<String?>? command,
-      Value<DateTime>? createdAt,
-      Value<DateTime>? updatedAt,
-      Value<DateTime?>? deletedAt,
+      Value<DateTime?>? localCreatedAt,
+      Value<DateTime?>? remoteCreatedAt,
+      Value<DateTime?>? localUpdatedAt,
+      Value<DateTime?>? remoteUpdatedAt,
+      Value<DateTime?>? localDeletedAt,
+      Value<DateTime?>? remoteDeletedAt,
       Value<String?>? userId,
       Value<bool>? pinned,
       Value<DateTime?>? pinnedAt,
@@ -2797,9 +3052,12 @@ class PinnedMessagesCompanion extends UpdateCompanion<PinnedMessageEntity> {
       showInChannel: showInChannel ?? this.showInChannel,
       shadowed: shadowed ?? this.shadowed,
       command: command ?? this.command,
-      createdAt: createdAt ?? this.createdAt,
-      updatedAt: updatedAt ?? this.updatedAt,
-      deletedAt: deletedAt ?? this.deletedAt,
+      localCreatedAt: localCreatedAt ?? this.localCreatedAt,
+      remoteCreatedAt: remoteCreatedAt ?? this.remoteCreatedAt,
+      localUpdatedAt: localUpdatedAt ?? this.localUpdatedAt,
+      remoteUpdatedAt: remoteUpdatedAt ?? this.remoteUpdatedAt,
+      localDeletedAt: localDeletedAt ?? this.localDeletedAt,
+      remoteDeletedAt: remoteDeletedAt ?? this.remoteDeletedAt,
       userId: userId ?? this.userId,
       pinned: pinned ?? this.pinned,
       pinnedAt: pinnedAt ?? this.pinnedAt,
@@ -2865,14 +3123,23 @@ class PinnedMessagesCompanion extends UpdateCompanion<PinnedMessageEntity> {
     if (command.present) {
       map['command'] = Variable<String>(command.value);
     }
-    if (createdAt.present) {
-      map['created_at'] = Variable<DateTime>(createdAt.value);
+    if (localCreatedAt.present) {
+      map['local_created_at'] = Variable<DateTime>(localCreatedAt.value);
     }
-    if (updatedAt.present) {
-      map['updated_at'] = Variable<DateTime>(updatedAt.value);
+    if (remoteCreatedAt.present) {
+      map['remote_created_at'] = Variable<DateTime>(remoteCreatedAt.value);
     }
-    if (deletedAt.present) {
-      map['deleted_at'] = Variable<DateTime>(deletedAt.value);
+    if (localUpdatedAt.present) {
+      map['local_updated_at'] = Variable<DateTime>(localUpdatedAt.value);
+    }
+    if (remoteUpdatedAt.present) {
+      map['remote_updated_at'] = Variable<DateTime>(remoteUpdatedAt.value);
+    }
+    if (localDeletedAt.present) {
+      map['local_deleted_at'] = Variable<DateTime>(localDeletedAt.value);
+    }
+    if (remoteDeletedAt.present) {
+      map['remote_deleted_at'] = Variable<DateTime>(remoteDeletedAt.value);
     }
     if (userId.present) {
       map['user_id'] = Variable<String>(userId.value);
@@ -2923,9 +3190,12 @@ class PinnedMessagesCompanion extends UpdateCompanion<PinnedMessageEntity> {
           ..write('showInChannel: $showInChannel, ')
           ..write('shadowed: $shadowed, ')
           ..write('command: $command, ')
-          ..write('createdAt: $createdAt, ')
-          ..write('updatedAt: $updatedAt, ')
-          ..write('deletedAt: $deletedAt, ')
+          ..write('localCreatedAt: $localCreatedAt, ')
+          ..write('remoteCreatedAt: $remoteCreatedAt, ')
+          ..write('localUpdatedAt: $localUpdatedAt, ')
+          ..write('remoteUpdatedAt: $remoteUpdatedAt, ')
+          ..write('localDeletedAt: $localDeletedAt, ')
+          ..write('remoteDeletedAt: $remoteDeletedAt, ')
           ..write('userId: $userId, ')
           ..write('pinned: $pinned, ')
           ..write('pinnedAt: $pinnedAt, ')
@@ -3060,7 +3330,7 @@ class $PinnedMessageReactionsTable extends PinnedMessageReactions
   }
 
   static TypeConverter<Map<String, Object?>, String> $converterextraData =
-      MapConverter<Object?>();
+      MapConverter();
   static TypeConverter<Map<String, Object?>?, String?> $converterextraDatan =
       NullAwareTypeConverter.wrap($converterextraData);
 }
@@ -3403,7 +3673,7 @@ class $ReactionsTable extends Reactions
   }
 
   static TypeConverter<Map<String, Object?>, String> $converterextraData =
-      MapConverter<Object?>();
+      MapConverter();
   static TypeConverter<Map<String, Object?>?, String?> $converterextraDatan =
       NullAwareTypeConverter.wrap($converterextraData);
 }
@@ -3790,7 +4060,7 @@ class $UsersTable extends Users with TableInfo<$UsersTable, UserEntity> {
   }
 
   static TypeConverter<Map<String, Object?>, String> $converterextraData =
-      MapConverter<Object?>();
+      MapConverter();
 }
 
 class UserEntity extends DataClass implements Insertable<UserEntity> {
