@@ -21,9 +21,13 @@ extension MessageEntityX on MessageEntity {
           final json = jsonDecode(it);
           return Attachment.fromData(json);
         }).toList(),
-        createdAt: createdAt,
         extraData: extraData ?? <String, Object>{},
-        updatedAt: updatedAt,
+        createdAt: remoteCreatedAt,
+        localCreatedAt: localCreatedAt,
+        updatedAt: remoteUpdatedAt,
+        localUpdatedAt: localUpdatedAt,
+        deletedAt: remoteDeletedAt,
+        localDeletedAt: localDeletedAt,
         id: id,
         type: type,
         status: status,
@@ -37,7 +41,6 @@ extension MessageEntityX on MessageEntity {
         showInChannel: showInChannel,
         text: messageText,
         user: user,
-        deletedAt: deletedAt,
         pinned: pinned,
         pinnedAt: pinnedAt,
         pinExpires: pinExpires,
@@ -59,7 +62,8 @@ extension MessageX on Message {
         parentId: parentId,
         quotedMessageId: quotedMessageId,
         command: command,
-        createdAt: createdAt,
+        remoteCreatedAt: remoteCreatedAt,
+        localCreatedAt: localCreatedAt,
         shadowed: shadowed,
         showInChannel: showInChannel,
         replyCount: replyCount,
@@ -67,10 +71,12 @@ extension MessageX on Message {
         reactionCounts: reactionCounts,
         mentionedUsers: mentionedUsers.map(jsonEncode).toList(),
         status: status,
-        updatedAt: updatedAt,
+        remoteUpdatedAt: remoteUpdatedAt,
+        localUpdatedAt: localUpdatedAt,
         extraData: extraData,
         userId: user?.id,
-        deletedAt: deletedAt,
+        remoteDeletedAt: remoteDeletedAt,
+        localDeletedAt: localDeletedAt,
         messageText: text,
         pinned: pinned,
         pinnedAt: pinnedAt,
