@@ -38,7 +38,8 @@ void main() {
       parentId: 'testParentId',
       quotedMessageId: quotedMessage.id,
       command: 'testCommand',
-      createdAt: DateTime.now(),
+      localCreatedAt: DateTime.now(),
+      remoteCreatedAt: DateTime.now().add(const Duration(seconds: 1)),
       shadowed: math.Random().nextBool(),
       showInChannel: math.Random().nextBool(),
       replyCount: 33,
@@ -52,10 +53,12 @@ void main() {
         jsonEncode(User(id: 'testuser')),
       ],
       status: MessageSendingStatus.sent,
-      updatedAt: DateTime.now(),
+      localUpdatedAt: DateTime.now(),
+      remoteUpdatedAt: DateTime.now().add(const Duration(seconds: 1)),
       extraData: {'extra_test_data': 'extraData'},
       userId: user.id,
-      deletedAt: DateTime.now(),
+      localDeletedAt: DateTime.now(),
+      remoteDeletedAt: DateTime.now().add(const Duration(seconds: 1)),
       messageText: 'Hello',
       pinned: true,
       pinExpires: DateTime.now().toUtc(),
@@ -81,7 +84,8 @@ void main() {
     expect(message.parentId, entity.parentId);
     expect(message.quotedMessageId, entity.quotedMessageId);
     expect(message.command, entity.command);
-    expect(message.createdAt, isSameDateAs(entity.createdAt));
+    expect(message.localCreatedAt, isSameDateAs(entity.localCreatedAt));
+    expect(message.remoteCreatedAt, isSameDateAs(entity.remoteCreatedAt));
     expect(message.shadowed, entity.shadowed);
     expect(message.showInChannel, entity.showInChannel);
     for (var i = 0; i < message.mentionedUsers.length; i++) {
@@ -93,10 +97,12 @@ void main() {
     expect(message.reactionScores, entity.reactionScores);
     expect(message.reactionCounts, entity.reactionCounts);
     expect(message.status, entity.status);
-    expect(message.updatedAt, isSameDateAs(entity.updatedAt));
+    expect(message.localUpdatedAt, isSameDateAs(entity.localUpdatedAt));
+    expect(message.remoteUpdatedAt, isSameDateAs(entity.remoteUpdatedAt));
     expect(message.extraData, entity.extraData);
     expect(message.user!.id, entity.userId);
-    expect(message.deletedAt, isSameDateAs(entity.deletedAt));
+    expect(message.localDeletedAt, isSameDateAs(entity.localDeletedAt));
+    expect(message.remoteDeletedAt, isSameDateAs(entity.remoteDeletedAt));
     expect(message.text, entity.messageText);
     expect(message.pinned, entity.pinned);
     expect(message.pinExpires, isSameDateAs(entity.pinExpires));
@@ -144,7 +150,8 @@ void main() {
       parentId: 'testParentId',
       quotedMessageId: quotedMessage.id,
       command: 'testCommand',
-      createdAt: DateTime.now(),
+      localCreatedAt: DateTime.now(),
+      createdAt: DateTime.now().add(const Duration(seconds: 1)),
       shadowed: math.Random().nextBool(),
       showInChannel: math.Random().nextBool(),
       replyCount: 33,
@@ -157,10 +164,12 @@ void main() {
         (prev, curr) =>
             prev?..update(curr.type, (value) => value + 1, ifAbsent: () => 1),
       ),
-      updatedAt: DateTime.now(),
+      localUpdatedAt: DateTime.now(),
+      updatedAt: DateTime.now().add(const Duration(seconds: 1)),
       extraData: const {'extra_test_data': 'extraData'},
       user: user,
-      deletedAt: DateTime.now(),
+      localDeletedAt: DateTime.now(),
+      deletedAt: DateTime.now().add(const Duration(seconds: 1)),
       text: 'Hello',
       pinned: true,
       pinExpires: DateTime.now(),
@@ -179,7 +188,8 @@ void main() {
     expect(entity.parentId, message.parentId);
     expect(entity.quotedMessageId, message.quotedMessageId);
     expect(entity.command, message.command);
-    expect(entity.createdAt, isSameDateAs(message.createdAt));
+    expect(entity.localCreatedAt, isSameDateAs(message.localCreatedAt));
+    expect(entity.remoteCreatedAt, isSameDateAs(message.remoteCreatedAt));
     expect(entity.shadowed, message.shadowed);
     expect(entity.showInChannel, message.showInChannel);
     expect(entity.replyCount, message.replyCount);
@@ -188,10 +198,12 @@ void main() {
     expect(entity.reactionScores, message.reactionScores);
     expect(entity.reactionCounts, message.reactionCounts);
     expect(entity.status, message.status);
-    expect(entity.updatedAt, isSameDateAs(message.updatedAt));
+    expect(entity.localUpdatedAt, isSameDateAs(message.localUpdatedAt));
+    expect(entity.remoteUpdatedAt, isSameDateAs(message.remoteUpdatedAt));
     expect(entity.extraData, message.extraData);
     expect(entity.userId, message.user!.id);
-    expect(entity.deletedAt, isSameDateAs(message.deletedAt));
+    expect(entity.localDeletedAt, isSameDateAs(message.localDeletedAt));
+    expect(entity.remoteDeletedAt, isSameDateAs(message.remoteDeletedAt));
     expect(entity.messageText, message.text);
     expect(entity.pinned, message.pinned);
     expect(entity.pinExpires, isSameDateAs(message.pinExpires));
